server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.9
PyJWT==2.8.0

